requirements.txt
# IDnow AutoIdent - Python Dependencies

# Core Appium dependencies
Appium-Python-Client
selenium

# Test framework
pytest
pytest-html

# Utilities
python-dotenv
